requirements.txt
fastapi
uvicorn
sentence-transformers
torch
transformers
accelerate
diskcache
python-dotenv
numpy
